adk-agents/memory_agents/requirements.txt
streamlit>=1.47.1
python-dotenv>=1.1.1
google-adk>=1.9.0
asyncio 